The current algorithm
The initial program. This current version is very slow. Due to making sure the program is broad enough for type aheads, the program has to allow a multitude of words past the "type ahead allowance" point. Next steps will be to enforce more on that position, and to purposefully stop following certain paths earlier on.